Add optional static middleware in test server
